% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bscard_template.R
\name{bscard_template}
\alias{bscard_template}
\title{Template for bscard}
\usage{
bscard_template(width = NULL, height = NULL, elementId = NULL)
}
\value{
shiny.tag
}
\description{
Template for bscard
}
