% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bscard_template.R
\name{task_card_template}
\alias{task_card_template}
\title{Template for task card}
\usage{
task_card_template(
  elementId = NULL,
  assignee = NULL,
  reviewer = NULL,
  date_due = NULL,
  process = NULL
)
}
\value{
shiny.tag
}
\description{
Template for task card
}
