% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bscard.R
\name{task_card}
\alias{task_card}
\title{Task card with custom elements}
\usage{
task_card(elementId = NULL)
}
\arguments{
\item{elementId}{id}
}
\value{
shiny.tag
}
\description{
Task card with custom elements
}
